requirements.txt
psycopg2-binary==2.7.5
aniso8601==3.0.2
astroid==2.0.4
bcrypt==3.1.4
cffi==1.11.5
click==6.7
colorama==0.3.9
coverage==4.5.1
Flask==1.0.2
Flask-API==1.0
Flask-Bcrypt==0.7.1
Flask-JWT==0.3.2
flask-restplus==0.11.0
Flask-Script==2.0.6
gunicorn==19.9.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
nose==1.3.7
psycopg2==2.7.5
pycparser==2.18
PyJWT==1.4.2
pylint==2.1.1
python-dotenv==0.9.1
pytz==2018.5
six==1.11.0
typed-ast==1.1.0
Werkzeug==0.14.1
wrapt==1.10.11
